cast LeptonsPerCell to double in some division scenarios

+++ src/Ext/Script/Body.cpp
@@ -432,7 +432,7 @@ void ScriptExt::Mission_Gather_NearTheLeader(TeamClass* pTeam, int countdown)
 		}
 		else
 		{
-			closeEnough = RulesClass::Instance->CloseEnough / Unsorted::LeptonsPerCell;
+			closeEnough = RulesClass::Instance->CloseEnough / (double)Unsorted::LeptonsPerCell;
 		}
 
 		// The leader should stay calm & be the group's center
@@ -470,7 +470,7 @@ void ScriptExt::Mission_Gather_NearTheLeader(TeamClass* pTeam, int countdown)
 
 				nUnits++;
 
-				if ((pUnit->DistanceFrom(pLeaderUnit->GetCell()) / Unsorted::LeptonsPerCell) > closeEnough)
+				if ((pUnit->DistanceFrom(pLeaderUnit->GetCell()) / (double)Unsorted::LeptonsPerCell) > closeEnough)
 				{
 					// Leader's location is too far from me. Regroup
 					if (pUnit->Destination != pLeaderUnit)
@@ -677,7 +677,7 @@ void ScriptExt::SetCloseEnoughDistance(TeamClass* pTeam, double distance)
 		pTeamData->CloseEnough = distance;
 
 	if (distance <= 0)
-		pTeamData->CloseEnough = RulesClass::Instance->CloseEnough / Unsorted::LeptonsPerCell;
+		pTeamData->CloseEnough = RulesClass::Instance->CloseEnough / (double)Unsorted::LeptonsPerCell;
 
 	// This action finished
 	pTeam->StepCompleted = true;
@@ -713,7 +713,7 @@ bool ScriptExt::MoveMissionEndStatus(TeamClass* pTeam, TechnoClass* pFocus, Foot
 	if (!pFocus || mode < 0 || (mode != 2 && mode != 1 && !pLeader))
 		return false;
 
-	double closeEnough = RulesClass::Instance->CloseEnough / Unsorted::LeptonsPerCell;
+	double closeEnough = RulesClass::Instance->CloseEnough / (double)Unsorted::LeptonsPerCell;
 	auto const pTeamData = TeamExt::ExtMap.Find(pTeam);
 
 	if (pTeamData->CloseEnough > 0)
@@ -734,7 +734,7 @@ bool ScriptExt::MoveMissionEndStatus(TeamClass* pTeam, TechnoClass* pFocus, Foot
 			if (mode == 2)
 			{
 				// Default mode: all members in range
-				if ((pUnit->DistanceFrom(pFocus->GetCell()) / Unsorted::LeptonsPerCell) > closeEnough)
+				if ((pUnit->DistanceFrom(pFocus->GetCell()) / (double)Unsorted::LeptonsPerCell) > closeEnough)
 				{
 					bForceNextAction = false;
 
@@ -758,7 +758,7 @@ bool ScriptExt::MoveMissionEndStatus(TeamClass* pTeam, TechnoClass* pFocus, Foot
 				if (mode == 1)
 				{
 					// Any member in range
-					if ((pUnit->DistanceFrom(pFocus->GetCell()) / Unsorted::LeptonsPerCell) > closeEnough)
+					if ((pUnit->DistanceFrom(pFocus->GetCell()) / (double)Unsorted::LeptonsPerCell) > closeEnough)
 					{
 						if (pUnit->WhatAmI() == AbstractType::Aircraft && pUnit->Ammo > 0)
 							pUnit->QueueMission(Mission::Move, false);
@@ -782,7 +782,7 @@ bool ScriptExt::MoveMissionEndStatus(TeamClass* pTeam, TechnoClass* pFocus, Foot
 					// All other cases: Team Leader mode in range
 					if (pLeader)
 					{
-						if ((pUnit->DistanceFrom(pFocus->GetCell()) / Unsorted::LeptonsPerCell) > closeEnough)
+						if ((pUnit->DistanceFrom(pFocus->GetCell()) / (double)Unsorted::LeptonsPerCell) > closeEnough)
 						{
 							if (pUnit->WhatAmI() == AbstractType::Aircraft && pUnit->Ammo > 0)
 								pUnit->QueueMission(Mission::Move, false);
